[dnn-library] [SW-2026] Moved the dnn_lib private headers to include/private
The include directory had too many header files that were not released
to the staging include directory. Those headers were moved to
include/private to ease the work of understanding what functions
are actually provided by the library.

While I was at it I saw that the AutoGenInstan.def.orig file was a stale
file with an old copy of AutoGenInstan.def which should have been deleted
when the new version was introduced. This stale file is removed for
avoiding confusion.