[dnn-library] [SW-2309] Create logic operations with an immediate operand in the lib
This commit implements in the lib the andi, ori and xori operations. These operations take a tensor and
an immediate value and apply the logic operation to each element of the tensor using the immediate value.
The implementation is done sequentially, using only one minion.

For example:

dst[i] = src[i] & imm