[dnn-library] [SW-5213] Fixed the scratch recompute for the DNN Lib TopK operator
The scratch recompute was broken in two ways:

1) Was applying the quicksort on every iteration, irrespective of
   whether the content had changed at all, which surely had a negative
   impact on performance.

2) The engineers having written the code or applied fixes did not
   get right the condition for deciding when to recompute.

Regarding the second point the git history shows two steps towards
the right solution:

- First step, by the code author, who thought of recomputing every time
  the coordinate with the highest index became zero. While this works
  for the case of having just one minion it fails when more than one
  is deployed.

- Second step, by Guillem, who realized the recompute was not
  happening often enough, and possibly it was missing for some test in
  which the scratch was not recomputed between batch elements.

The second step brings us closer to the final solution applied with the
current patch, which hopefully is complete (until more topK tests are
available for uncovering bugs). With the current patch the recompute is
trigrered every time that all the coordinates but the last change.

For understanding why the proposed current fix works let us assme
that a row is composed of columns and the column index is the innermost
dimension. The first time a minion is deployed on a row it recomputes
the topK for the row into the scratch. On subquent minion redeployments
on the same row the scratch content is reused rather than recomputed.

The case of switching from one batch to the next that Guillem ad-hoc
fixed is just one of the many ways in which all the coordinates but
the last one will stop matching with the coordinates from the
previous processed element.