[dnn-library] [SW-5323] Fixed a bug on the fwdLibAvgPoolInstThreaded arguments order
The graph-optimizer was forcing the non-threaded average pooling
implementation from DNN Lib, because the threaded implementation was
seen to cause an issue when running with the new countIncludePads
parameter set to false.

The issue has now been investigated and we found out that there
was an actual bug due to the arguments for fwdLibAvgPoolInstThreaded
being wrongly ordered on the function definition.

What this patch does is to:

- Fix the order of the arguments on the fwdLibAvdPoolInstThreaded
  definition.

- Remove the bug workaround that was forcing the non-threaded variant
  to be used instead of the threaded one.