[dnn-library] [SW-4415] A better way of implementing round-towards-zero
For the convertion from FloatTy to Int32ITy we were implementing
rounding towards zero (rtz) by means of round-to-nearest with half-way
to maximum magnitude (rmm) as:

rtz(x) = rmm(x - sign(x)*0.5f)

Where the the "sign" function returns +1 when x is positive and -1 when
negative.

While the trick is clever and cheap, we actually have the RTZ rounding
mode in fcvt.pw.ps, which should be used instead and it is unbeatable,
since it allows to do the rounding with only one instruction, as opposed
to 4 which were needed before this patch.