[dnn-library] [SW-5652] Unify the vectorized SparseToDenseOp variants
Provide just one general implementation of SparseToDenseOp, rather than
repeating the function structure as many times as supported types. This
move is useful for adding support for new types (like Int64ITy which is
missing and will soon be added).

There is no code efficiency penalty paid for the unification, as the
type specifics are handled at built time with "if constexpr" constructs,
which leads to efficient code, on par with the one previously generated.
This fact was double checked by looking at the generated assembly.