[dnn-library] [SW-2630] Encapsulate fmadd in an inline function
While I was at it I also refactored the doDequantize function for using
the newly created function. In the process I fixed a latent bug:
the destination register was declared with "+&f" when it should have
been rather declared as "=&f". Luckily the code was still not used and
caused no issue.