[dnn-library] [SW-2630] Patch NaN result on quantized case of fwdElementDiv
On the quantized fwdElementDiv case it has been found that the result
of "lhs * rcp(rhs) + destOffset" can be NaN. Then when converting from
float to int32_t we have that RISC-V and x86 produce different values:
RISC-V produces 0x7fffffff whereas x86 produces 0x80000000.

We already had a patch for matching the case when the expression
result is +Inf, which suffered from a similar issue. The present
patch simply completes the existing workaround for handling the NaN
case.

Overall, 6 extra vector instructions are added for handling the NaN
and +Inf cases. One should be able to strip this overhead when
building DNN Lib with a "fast-math" style flag.