[dnn-library] [SW-4998] Refactored the FloatTy to Int32ITy converts
The code converting from FloatTy to Int32ITy was duplicated. Now, on the
other hand, the code is encapsulated on a templated function that allows
the caller to freely pick the rounding mode at compile time.

We have also made optional the check for signalling NaN, given that
depending on the context from which the function is called, it may be
assured that any NaN is quite rather than signalling. For example, if
we are converting an intermediate result rather than a user input, it
can not be a singalling NaN, and therefore there is no need to check
for that case.

This refactor was needed for producing a fix for SW-4998, which is due
to wrong rounding.