[dnn-library] [SW-5024] Fixed accumulators initialization on convolution vector code
The vector code for FloatTy and Float16Ty convoluntion was
clearing the accumulators before setting the vector mask.
This was wrong, as dependending on the mask content carried over
from the previous operator on the nework, all the accumulators
would be cleared or not, with random side effects.

This was a latent bug that was only observed on the FP32
mnist model after some urelated changes were made on MaxPool,
but may have affected other more complex models.

The way of reproducing the issue was to apply the incoming patches for
MaxPool and running the "vip models" CI test suite as:

  make test-models-ci-vip-etrt-devfw-sysemu -j20

Now it is well understood that the incoming MaxPool patches uncovered
the bug by using some vector code that set the mask to 0x1, rather than
not touching the mask, as the previous MaxPool implementation did.

The bug has been fixed by doing a sensible reordering of the
instructions on the convolution code.