[dnn-library] [SW-5846] Support destination buffers with different padding for LRN
The LocalResponseNormalization has two output tensors that may have
different padding, however our code made the assumption that the padding
was exactly the same. Now the code has been improved to removing this
assumption.

We still need to provide a mechanism for the second tensor to not to
suffer from a cache coherency issue derived from the fact that it is
not guaranteed that at most only minion writes to a given cache line.
We may need to address this issue as part of SW-6478.