[dnn-library] [SW-6517] Do not select the vectorized SoftMax for unsupported cases
The vectorized implementation is known to suffer from a cache coherency
hazard. That hazard is not trivially fixed on the current implementation
without affecting the performance and making the code overly complex.

The vectorized implementation should be avoided for cases other than:
 - having two dimensions
 - an start address aligned to cache line
 - a stride multiple of the cache line

A better implementation is possible that is still performant but does not
assume such a strict alignment. That brand new implementation would
use an aligned scratch buffer with one line per row:

1) Divide the number of rows by the number of available minions. Minions
   with a numeric id above the remainder of rows will do an extra row,
   so all the rows are processed. For each assigned row a minion stores
   the per-row max value on the corresponding cache line.

2) Do a second reduce pass with the same minion deployment, but
   calculating the sum over "row" of "exp(input(batch,row) - max(row))"
   The result is also stored in the corresponding cache line.

3) Use the partitionLoop infrastructure for outputting the expression:

   output(batch,row) := exp((input(batch,row)) - max(row)) / sum(row)