[dnn-library] [SW-6517] Updated libManager.xlsl and regenerated the inlining stubs
On the "instances" tab from libManager.xlsl, changed the SoftMax row
so it contains only three types: FloatTy, Float16Ty and BFloat16.

Those are the only types supported by the SoftMax implementation from
the Glow interpreter and are the only ones for which should be
instanciated. Moreover, our implementation did not work for types other
that the three here mentioned.