[dnn-library] [SW-10667] Fixed the expected padding array layout for 3D convolution
Some tests for 3D convolution using padding were failing. The reason
for the issue was that our implementation assumed the array of padding
values was in NTLFBR format, whereas Glow provides the padding array as
NFTBLR. This was fixed.

While I was at it I also cleaned up some commented out code and
redundant or stale comments.