[dnn-library] Refactor and document the template arguments on ConvolutionInst.h
Unified the template parameters N and KN, that were always deducted
by compiler as being the same value. With this patch both parameters
become just N.

Doxygen documentation was updated where stale references to src1Type,
src2Type and dstType where present, and needed to be replaced for
dstElK and biasElK.

The doxygen style guide favouring triple slash was not followed and
needed to be addressed.