[dnn-library] Fix for regression on ConvertTo test suite
Handling of non-finite floating point numbers on convertion to
uint32_t can be configured in LoadStore.h for matching x86 or not,
and by default it does try to match.

A previous patch changed the default behaviour for not trying to
match, leading to some tests on the ConvertTo failing. This patch
restores the previous match default behaviour, making the
regressed tests to pass again.

Testing:

I ran the random suite for Convert, seeing that the failures got
fixed.

make test-op-dnn-lib-ConvertTo-tests-random-etrt-devfw-sysemu -j10