[dnn-library] [SW-11228] New rounding mode supported on convert
On doQuantize and convert template functions, now we allow rounding like
nearbyintf with the default rounding mode on GCC and Clang (RNE), followed by
a cast to  int32_t. This was done as a result of having reviewed the current
doQuantize interface and finding out that we were not supporting
this kind of rounding, which is used by Glow in some cases.