[dnn-library] [SW-11122] New simpler maxplat implementation
The previous version was converting the source tensor to float
for applying the max operation between float elements, but a
better approach was possible by instead converting the splat
value to the source tensor type and operating with that type.

While I was at it I also did an experiment of reworking the
partitionLoop infrastructure so the caller can avoid passing
the compute operation as a function pointer, potentially messing
with the inlining of the function. The new alternative kernel
dispatching mechanism uses a preprocessor parameter for passing
the compute function name.