[dnn-library] [SW-11876] Fix for colliding FLB counters
The fast local barriers were always using the same counter (0),
which led to groups of minions working on different things to use
the same counter and messing each other's synchronization.

A simple mechanism for assigning counters to different minion
groups was introduced. Now, each group uses the local id of the
first minion in the group, which uniquely identifies the minion
group and fixes the issue.