[dnn-library] FFT: align the auxillary stack allocations to cache lines
This patch fixes a few read coherency hazards we saw. The hazards
were found to be due to the temporary buffers not being aligned and
spanning one more line than it was calculated when assuming alignment.

In order to fix the problem I have added alignment to all the
allocations on the auxillary stack, so the assumption is correct.