[dnn-library] FFT: Fixed a race condition on intra-row tiling
All the minions assigned to a given column were writting the final
result to the output tensor, when just the first minion on the
tiling group should do it.

While I was at it I also did a clean up by replacing an evict plus
normal load with atomic loads, which seems like a cleaner solution
and makes the code more understandable.

Finally, a barrier was added at the end of fft2d_threaded to ensure
all the pending minion work is finished and the destination tensor
is ready in DDR memory.