[dnn-library] [SW-12585] Create a branch in sw-platform to execute the dnnLibrary CI
This is needed to avoid unexpected auto-cancels in sw-platform.

The new feature has been implemented by including the file "sw-platform-trigger.yaml" from gitlab-ci-common. This file has
the jobs to create a branch in sw-platform and to trigger the CI there using the newly created branch.

Apart from that, the "check-on-top-of-master.yaml" job has also been included in the Neuralizer's CI to ensure that all the
CIs start with a branch which is on top of master.