[dnn-library] [SW-13430] Fixed a bug on the FFT reduce code
The vector code was always used and it assumes halfSize is a multiple of 8.
However, for cases when halfSize is not a multiple of 8 the previous
scalar code should be used instead. The code has been fixed accordingly.

This fixes an issue seen with sizes smaller than 8 pixels wide or high.