[gp-sdk] Make the section with crt0 code match the first section on linker.ld
The section with the crt0 code is intended to be placed a the base
address where the .elf is loaded. However, because it is called
.text.startup and the linker script assumes it is called .text.init,
the linker does not always place crt0 as the first section.

It has been observed on programs linked with the script prior to
applying this patch that the crt0 section was indeed misplaced.

This patch fixes the issue by renaming the code section in crt0.s
as .text.init

With the patch I also remove a workaround for the debugger setup
(that was ignoring the first 4096 after the base address RVA). This is
done as per request (graceful advise) of Rafael Jimenez.