[gp-sdk] Set the base RVA for .elf start to zero (rather than 1)
Setting it to 1 was a hack that Grigorios introduced for avoiding
a binutils bug, where the assembler was unable to generate an jump
with an immediate address being zero that could later be fixed up
during the .elf load. The bug had to do with the instructions sequence
using the x0 RISC-V register rather than an actual immediate (or
something along those lines I can not accurately remember because this
was a conversation I overheard four years ago).

Setting the value to 1, did not trigger the usage of x0, and was also
zero because the address was rounded down to a multiple of 4 because of
jump addresses being aligned to 32 bits.

Whatever the bug was, it seems we are not hit, and the programs are
loading and executing well without this horrible hack (could be our
binutils got an update in the last four years). The customers should not
get this hack.

Note:

This fix needs to be ported to neuralizer.