[gp-sdk] Extra arguments on device/CMakeLists.txt for building custom kernels
Users of this CMakeLists.txt can provide an additional subdirectory by passing build arguments,
as opposed to modifying this CMakeLists.txt. This CMake configuration mechanism is intended for
consumers of the GP-SDK to build custom kernels leveraging the same build recipe from the
internal GP-SDK tests. The accepted parameters are:

 - CUSTOM_KERNELS_SRC_DIR: directory where the sources and the corresponding CMakeLists.txt
   are located. It can be located out of this tree.

 - CUSTOM_KERNELS_BIN_DIR: directory where the binary files will be stored. If the directory
   does not exist it gets created.

If one such pair of arguments are provided for building a custom kernel, they are handled at this stage.

This mechanism has been sucessfully tested for building the pagerank.elf
kernel (from the pagerank sample). The change will allow the samples
repo to consume the minion HAL functions, without substantially changing
the current project structure or release method.