[dnn-library] Fixed an alignof(void) statement that did not build with LLVM
A recent change on dnnLib used the alignof operator on a template
that is passed the void type.

The template instantiation led to the expression alignof(void) being
evaluated. The C++ standard states that a compiler should fail to
compile in this case, but GCC did actually build. When I trided LLVM,
the code actually failed to compile and needs to fixed.