[dnn-library] [SW-11312] Added explicit zero offset zero to {flw|fsw}.ps instances
For the flw.ps and fsw.ps memory access instructions, clang expects the
memory operand (second) to have an emplicit offset, rather than having
it omitted and assuming zero.

I have added the expected offset in all the cases when it was missing.
Some instances did not need updating because the memory operand was
marked with the "m" modifier (standing for "memory"). For operands
marked with "m" the compiler adds the zero offset by transforming an
operand like "rs" into "0(rs)". This behaviour of the "m" modifier is
consistent between gcc and clang, with both of them applying the same
substitution.

While I was at it, some memory access instructions with offset zero
expressed in hexadecimal ("0x0") got their offset expressed in decimal
format ("0"). IMO, this makes the code more readable, given that
a zero is written the same irrespective of the numeric base. The "0x"
prefix is just visual redundancy and can be avoided.