[dnn-library] [SW-16750] Make fgh.ps occurrences to follow the official format
The official format is "fgh.ps fd, fs1(rs2)" as opposed to the format
that was used in the EmbeddingBagInst implementation which was
"fgh.ps fd, fs1, rs2". The code has been fixed, because that instruction
format is not documented and does not need to be supported.

While I was at it I also fixed the same issue for other instructions
that share the same or similar format with the second operand using
two registers for addressing a memory address. The reviewed instructions
are:

- fgw.ps
- fgh.ps
- fgb.ps
- fscw.ps
- fsch.ps
- fscb.ps
- fg32b.ps
- fg32h.ps
- fg32w.ps
- fsc32b.ps
- fsc32h.ps
- fsc32w.ps