[dnn-library] [SW-16192] Fix creation of tensor iterator from offset pointing to padding.
The HandleIterator for Lib tensors is used to iterate over valid elements
(i.e., excluding padding). Sometimes, an iterator is created from an
offset and that would point to a padding element, so the iterator must
be moved to the next valid element in the tensor. The existing code
only worked for padding at the last dimension and now it is
generalized to work when the padding occurs at any given dimension.