[dnn-library] [SW-11425] Declare vector mask register usages for clang in include/Float16.h
On ETSoC1, when writing inline assembly for the clang compiler, the programmer
must declare the usage of mask register operands through constraints as he
does for any other register class.

For example, if a maskable vector instruction is used, then the m0
register is being used and it must be declared as an input (with the
ETSoC1-specific "M" modifier on the corresponding constraint). If vector
mask registers other than zero are used as operands, the also needs to
be declared (using the ETSoC1-specific "N" modifier).

Did the needed changes and in addition to that I removed:

- The usage of the mask_set function for clang, where m0 is now set with the
  aforementioned "M" operand.

- Two functions that were dead code and did not need to be maintained
  anymore.