[dnn-library] Forked core files from "./internal/" for supporting vector data types
The "./internal/" directory contains headers reused by all the operators,
which is a core of functionality that needs to be migrated to vector
datatypes first, before its depending operator implementations on
"include/inlining" are migrated.

In order to make the migration smooth the plan is to:

- Duplicate the core files from internal into files with the "2" suffix,
  as done in this patch.

- Migrate to vector data types the forked files and merge in one patch.

- In an orthogonal way port all the operators, merging one individual
  patch for every operator.

This strategy is better than, for example, migrating the files in the
"./internal/" folder and all their depending operators in just one patch
that takes weeks to produce.

The proposed incremental approach leads to an easy test patches sequence.
It also makes the migration work easy to split between several engineers.