[dnn-library] Removed the unused inlining/GenericHeaders.h
This file was including inlining/inlining.h so, indirectly, all the
operators were parsed on every compilation unit. Recently we replaced
references to GenericHeaders.h with references to inlining.h, removing
one level of indirection to make the dependency and the scope of the
include more clear to dnnLib maintainers.

The overall effect on the builds is neutral: for inlining we included
everything under the sun before this change and so we do now.