[dnn-library] [SW-16749] Fixed out of bounds access on AvgPoolInst.h
The AvgPoolInst operator handles 4D tensors (NHWC), but uses dim_array_t
that is a static array holding up to 6 coordinates. Then when passing
a dim_array_t to getNonPaddingCoordinates the number dimensions was
specificied as ARRAY_SIZE(coord) which is 6 (capacity) as opposed to the
actual array occupancy which is 4.

This patch fixes two tests from the dnnLib regression used for testing
clang and invoked as:

make test-op-dnn-lib-TestLaunchers-tests-minimal-ci -j50

It has been very lucky to get this test failing when switching
compilers, because this bug could have remained hidden way longer.