[dnn-library] [SW-18949] Fix argmax null pointer dereferencing in maxpooling operator
It has been found that the MNIST model fails with builds using clang as
the default compiler for neuralizer.

Investigation of the issue led to the finding that the maxpooling
implementation in dnnLib did not correctly handle the case when the
argmax output tensor is not used.

For supporting the case in which the argmax output tensor is not needed,
the maxPoolImplThreaded implementation of the maxpooling operator accepts
nullptr for the out2T parameter (argmax tensor). Incidentally, one model
that uses maxpooling in this way is MNIST.

Prior to this patch the maxpooling operator was simply dereferencing the
argmax pointer without checking whether it was nullptr. This patch adds
the needed checks, fixing the MNIST issue spottted with the clang builds.