[dnn-library] [SW-18949] Simplified the interface of maxPoolImplThreaded
Whether the argmax tensor should be used by the maxPoolImplThreaded
function was redundantly signalled by the caller through two
correlated mechanisms:

- passing a bool parameter
- passing a pointer to the argmax tensor, that needed to be nullptr
  if the bool was false.

I have simplified the interface so the bool parameter is no longer
passed. If the tensor is not used it is signalled by passing nullptr
as value for the pointer.