[dnn-library] [SW-20313] Implement interleaved DequantizeNBitsColumnBlocks.
The new implementation allows for both the source and the destination
matrices to have a number of rows interleaved. This interleave factor
must be the same for the source and the destination and must divide
the number of elements inside a block. In that case, the quantization
offset and scale corresponding to a matrix position must be chosen
taking into account how the matrix dimensions are altered by the
interleaving transformation.

The current implementations mix coordinates and offsets by means of
pitches. When a tensor has interleaving, Neuralizer sends the
parameters as follows:
- The tensor dimension sizes are the original ones.
- The tensor pitches are the original ones multiplied by the
  interleave factor.
In terms of the implementation of the operator, it makes more sense
to multiply/divide the dimension sizes by the interleave factor and
to have this change reflected in only one of the two pitches.